Move tgl to tgl subdirectory, and use target_include_directories
